Refactored the main class – functionality moved to MenuNavigator, ProgramRunner, UserSessionHandler. Made ValueValidator a utility class. Reworked the welcome message display.
