+++ CoworkingSpaceReservation/src/main/java/org/example/controller/ValueValidator.java
@@ -1,5 +1,6 @@
 package org.example.controller;
 
+import lombok.experimental.UtilityClass;
 import org.example.model.Admin;
 import org.example.model.SpaceType;
 import org.example.model.User;
@@ -9,13 +10,14 @@ import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Arrays;
 
+@UtilityClass
 public class ValueValidator {
     private static final int MIN_ALLOWED_VALUE = 1;
     private static final int MAX_ALLOWED_FOR_MAIN_MENU = 3;
     private static final int MAX_ALLOWED_FOR_CUSTOMER = 5;
     private static final int MAX_ALLOWED_FOR_ADMIN = 4;
 
-    public boolean checkValue(String message) {
+    public static boolean checkValue(String message) {
 
         if (message == null || message.length() != 1) return false;
 
@@ -27,7 +29,7 @@ public class ValueValidator {
         }
     }
 
-    public boolean checkValue(User currentUser, String message) {
+    public static boolean checkValue(User currentUser, String message) {
         if (message == null || message.length() != 1) return false;
 
         try {
@@ -40,7 +42,7 @@ public class ValueValidator {
         }
     }
 
-    public boolean checkPrice(String message) {
+    public static boolean checkPrice(String message) {
         try {
             return Double.parseDouble(message) > 0;
         } catch (NumberFormatException e) {
@@ -48,7 +50,7 @@ public class ValueValidator {
         }
     }
 
-    public boolean checkType(String message) {
+    public static boolean checkType(String message) {
         if (message == null || message.isEmpty()) {
             return false;
         }
@@ -56,7 +58,7 @@ public class ValueValidator {
                 .anyMatch(type -> type.getDisplayName().equalsIgnoreCase(message));
     }
 
-    public boolean checkIdValue(String message) {
+    public static boolean checkIdValue(String message) {
         if (message == null || message.isEmpty()) return false;
         try {
             Long.parseLong(message);
@@ -66,7 +68,7 @@ public class ValueValidator {
         }
     }
 
-    public boolean checkDataType(String message, DateTimeFormatter formatter) {
+    public static boolean checkDataType(String message, DateTimeFormatter formatter) {
 
         try {
             LocalDateTime inputDateTime = LocalDateTime.parse(message, formatter);
@@ -80,13 +82,13 @@ public class ValueValidator {
         }
     }
 
-    private boolean isBeforeCurrentTime(LocalDateTime inputDateTime) {
+    private static boolean isBeforeCurrentTime(LocalDateTime inputDateTime) {
 
         LocalDateTime currentDateTime = LocalDateTime.now();
         return inputDateTime.isBefore(currentDateTime);
     }
 
-    public boolean checkDataTypeAndRange(LocalDateTime startTime, String message, DateTimeFormatter formatter) {
+    public static boolean checkDataTypeAndRange(LocalDateTime startTime, String message, DateTimeFormatter formatter) {
         try {
             LocalDateTime parsedDate = LocalDateTime.parse(message, formatter);
             return parsedDate.isAfter(startTime);
@@ -95,7 +97,7 @@ public class ValueValidator {
         }
     }
 
-    public boolean checkLogin(String login) {
+    public static boolean checkLogin(String login) {
         if (login == null || login.length() < 5) {
             return false;
         }
